+++ src/utils/mod.rs
@@ -3,3 +3,4 @@ pub mod pack;
 pub mod cli;
 pub mod filesystem;
 pub mod chunks;
+pub mod check_mcbe_install;